SCICD-216: Modify the Git Hash for All Products
Update the git hash to the most recent commit on the branch for all products
in LOCATION_DICT.  This is now common functionality when doing NCN
Personalization and building the COS image.

Create a stage for the customize_cos_compute_image method rather than doing
it as part of the image building.