SCICD-222: NCN Personalization Always Times Out
Modify the logic a bit for the wait_for_ncn_personalization method so that
if a node is in an error state it isn't waited for.

Make the waiting logic for wait_for_pod a little more accurate.  So, pull the
logic for getting all the pods out of the `while` loop.  This will reduce the
time complexity a bit.