SCICD-234 and SCICD-238
SCICD-234: Remove hardcoding of analytics branch
Use the product catalog to get the analytics information, and remove
the ANALYTICS_BRANCH usage.

SCICD-238: dvs reload stage crashes when pod completes
Catch an IndexError when checking the pod status.  Once a pod is
deleted, it will not be listed in a `kubectl get pods ...` command.

Miscellaneous:
--  In the unload_dvs_and_lnet method within ShastaUpdate.py,
    move all the analytics logic to the check_analytics_mount method.

--  Remove the `if host == "lemondrop-ncn-m001"` reference when determining
    whether or not to call dvs_reload_ncn.  Add a check for lustre mounts on
    the node.

--  In vars.py when calling `sat showrev ...`, use yaml instead of json.
    Older versions of sat do not support json format.