SCICD-243 Fix writing new options to config file
Fix a problem related to writing the config file when entries didn't
previously exist in the config file.  The issue was updating the
dictionary at the top level which was just overwriting the saved parser
data with the config file verbatim without honoring anything that was in
the parser but not in the config.  Change the function to update section
by section instead which preserves everything.