SCICD-281: Stage Tracking
Create a class for the stages, since it takes a lot of processing in the
main script.  This will also help encapsulate the stage processing.

Add functionality to display the status of the stages, so a user can tell
whether or not a stage hasn't been ran, was skipped, succeeded, or failed.