SCICD-293: Handle x.y.z Versioning
This isn't directly related, but was discovered while working on this jira.

Modify the behavior of the `--worker-nodes` argument so that it doesn't match
regular expressions, and allows a list as input.  This is to prevent
mistakes in the input format, which could result in running the
unload_dvs_and_lnet stage on the wrong node.