SCICD-323 reduce the required number of deployments
With the new 2.3 proceedure, deployments are no longer removed because
the unmount/unload can happen while the deployment is in place.  As
a result we don't need to worry about the minimum level of deployements
so much.  Still check for less than 2 since 2 is the default, so
if you are less than 2 something may be wrong.  Also, leave the warnings
in place to call attention to CPS deployment scaling.