SCICD-549: Running extra stages with '-r'
In lib/stages.py, fix the logic.  Run stages and specifying the '-b'/'-e'
are mutually exclusive.  If `--run-stages` was specified, set the
_stages array to that.

Also, fix/simplify the logic in Activity.py.  We can basically determine
if process-media was called by looking at the products variable.  raise an
error if the products list is empty.