SCICD-605-giving-up-messages
This generally fixes the problem where we get the error message,
"Giving up following the log for pod ...".  The main problem was that
we allowed the thread to continue after the pod had already exited.

If we still get this error, we might be able tighten up that logic a bit,
but this should fix the vast majority of cases.