SCICD-636: --mask-recipe-prods Fix
We need to check that config.args["mask_recipe_prod"] is defined before
checking it's type.

Also, we need to do a copy.deepcopy of the payload when assigning it to
patched_payload, so that when the "force" element is removed from
patched_payload, it's not also removed from payload.