SCICD-622: --media-dir/-m processing
Loosen the requirements on media dir.  Since it defaults to
${RBD_BASE_DIR}/${ACTIVITY_NAME}, it's never strictly required; but it
does need to exist for process-media and the bootprep stages.

This also resolves SCICD-616.  If an activity is not specified in the `ls`
subcommand, just list the stages and their description.

Also resolve SCICD-660 by moving the `configfile.write()` call back into
main.