SCICD-659: Add the Media Directory and Commands to activity_dict.yaml.
Add the media directory and commands to activity_dict.yaml.  Also, dump the
worflow info to $STATE_DIR/sessions/$SESSION_ID.  This is useful for
debugging.

The bulk of these changes relate to changing the definition of the state
function within the Activity class.  Two more arguments would have needed
to be added, so the definition was changed to use a dictionary rather than
listing each argument explicitly.

This PR also resolves the following:
SCICD-665: Stack Traces with No Arguments
We need to check if args has a func attribute prior to checking it's value.

SCICD-667: Argo Log Flat Files Truncated.
The code was breaking early and exiting.  We don't need the break,
because the individual threads will get cleaned up once the pod quits
running.  So All we really need to do is remove the break statement.

SCICD-662: Stack Trace when moving leader
When migrating pods, the pod won't be available for a few seconds,
and urllib3.exceptions.MaxRetryError will be raise.  Catch this exception
and continue to poll for POLL_LOGS seconds.