SCICD-619: Ctrl-C Problems
Using the threading module requires a lot of context-switching overhead
when switching between threads, and the CPU was always maxed out, so the
Ctrl-C's seemed unresponsive.  Switch to using the multithreading module.
This seems to keep the Ctrl-C's responsive.

Put a wrapper around `self.core.list_namespaced_pod` in PodLogs.py.
When spamming Ctrl-C's, it encounters network errors.  It's still possible
to hit network errors, but this makes it more resilient.  A user needs to
really try to hit the network errors now.

Start the PodLog threads with a nice value of 20.  This is as nice
as they can get.  Also, clean up entres in `self._running_subthreads` as
pods finish to reduce the memory usage.

With the change of threading model, it's not necessary to worry about the
reentry locks, so delete initialization of RLOCK from vars.py, and it's usage
from iuf and Activity.py.

In iuf, sleep for a fraction of a second while handling the Ctrl-C interrupts.
It saves a little processing if a user is holding down the key and
spamming them.