CASMINST-6518: Fix incorrect sat bootprep commands
Fix incorrect `sat bootprep` commands that are output by the `iuf` CLI
for debugging. The incorrect `--limit` option values and `--overwrite-*`
options were specified for the `prepare-images` stage. In addition,
there were redundant `cd` commands, and the indentation was not ideal
for readability. Indent each continuation of a single command an
additional 4 spaces to make it clear it is a continuation of the
previous line.

Add unit tests. They still need to be added to be executed in the
pipeline.

Test Description:
Only tested with unit tests so far.

I plan to take the RPM built by the pipeline, install it on a system and
execute the `update-cfs-config` and `prepare-images` stages against a
simple invalid `sat bootprep` file and verify that the correct debugging
commands are output.