CASMINST-6518: Execute unit tests in the pipeline
Add unit test execution to the build pipeline. This is done by adding a
stage that executes in a Python container provided by the CSM docker
registry. This stage executes in a Python 3.10 docker container just
like the stage which builds the RPM. Since this is in a docker
container, just install the required Python packages directly in the
container. Then execute tests with `nose2`.

Note that the coverage plugin is not working, but this is due to the
`sqlite3` package in the Python standard library being missing from the
container image. Once that is fixed, the coverage should start working.

It is also possible to run unit tests locally using a virtual
environment created with `pyenv`. In that case, unit tests can be
executed by simply running `nose2`.

Test Description:
Ran unit tests locally using `make unittest` without `pyenv` active. Ran
unit tests locally using `pyenv` virtualenv. Verified unit tests work in
the Jenkins pipeline.