CASMINST-6518: Constrain semver package version and add requests
The `semver` package has released version 3.0 since it was added to this
repository. The new major version deprecated `VersionInfo` in favor of
`Version`. Update the code accordingly, and add a constraint on the
version of this package to prevent future breakages.

In addition, this code has a top-level requirement on `requests`, so add
it to the `requirements.txt` file with reasonable constraints.

(cherry picked from commit 527f7119e7991c6cd38cabd876e486e066de9d1e)