SCICD-914: Stacktrace Originating in PodLogs.py
lib/PodLogs.py: When podlogs couldn't be listed, podlogs would throw an exception in
the list_namespaced_pod function.  Rather than abort the whole process
because we can't access the pods, just print a debug message that we're
giving up on following that pod.  This is how it's done in `follow_pod_logs`.

lib/Activity.py: In the get_workflow function, don't raise an error if we
can't get the workflow.  When only one stage is ran, or when after the last
stage of several stages are ran, there isn't another workflow to run.  When
this happens, iuf will exit 1 without any indication that there was an error,
except the debug message in the log.